feat: implement Identify method (#89)
This commit adds support for `IdentifyAsync`, which causes the client to
shut down its data source, switch contexts, and restart the data source.

Since it returns a future, users can call `wait()` to make it
synchronous. An alternative API would be to accept a completion handler.
That would imply executing user callbacks on a shared or separate
executor, and probably isn't as convenient for the common case of
blocking.

To enable safe shutdown of the Eventsource/Datasources, I had to
refactor them to use shared ownership.

This is because the eventsource client (or polling requester) has
completion handlers queued within asio that have pointers back to the
objects.

If we've shutdown and destroyed those objects, then the handlers will
cause undefined behavior when they execute and access the destroyed
object.

Now, we pass in `weak_ptr` of the Datasource into the Eventsource
client's callbacks, and lock the weak pointer before accessing any
members (such as the logger). So if a completion handler finally
executes long after the object is destroyed, it should be harmless.