feat: follow redirects in SSE client (#104)
This PR allows the SSE client to follow temporary/permanent redirects. 

Going from http to https is not going to work because the initial client
wouldn't have the ssl context, but this should cover the basic case.

I've also run this revision 164 times through contract tests and didn't
observe a crash yet.