feat: flag data model (#156)
This PR builds on #153, extending the deserialization to flag data. 

I've added to the `PARSE_FIELD` macro to handle the following cases:
1. Parse a field where the zero-value is a valid part of the domain:
`PARSE_FIELD`. In other words, if the field is omitted or null, it will
be filled in with a default value. Most common scenario.
2. Like (1), but supply a default value in case the field is null or
omitted: `PARSE_FIELD_DEFAULT`.
3. Parse a field that doesn't have a valid default value, and so is
represented as an `optional`: `PARSE_CONDITIONAL_FIELD`. For example,
the value `""` wouldn't be a valid default for a unique key.
4. Parse a field that must be present in the JSON data:
`PARSE_REQUIRED_FIELD`. Usage of this should be rare; the test would be
if the SDK is totally unable to function without this.

To achieve this, I've made the `tag_invoke`s operate on
`expected<optional<T>, JsonError`'s, instead of `expected<T, JsonError`.

We could potentially revert that (and instead add a new enum case to
`JsonError`), but I figure `optional` is a good representation of
null/omitted. I think I need to actually try removing it to determine if
it results in a better API.

One thing lacking is better error messages - we use kSchemaFailure
everywhere but we now have the ability to report exactly what was
missing/wrong type, etc.