fix: refactor Variation methods for correctness & clarity (#203)
This significantly refactors the `Variation` internal calls to be much
easier to reason about. In the process, it fixes some broken behavior
which is now reflected in the suppressions list.

It also refactors event generation to use the
`EventScope`/`EventFactory` pattern found in other SDKs, such as Rust
and Go.