fix: initialization of LDFlagListener  (#218)
This commit moves the definition and implementation of `LDFlagListener`
into `common`, so it can be easily shared with the server.

Additionally, it fixes a bug in the specification of the
`LDFlagListener_Init` method. It previously took `LDFlagListener` by
value; now it takes a pointer.

The impact would be if a user forgot to initialize the struct, then
the member(s) would be undefined.