fix: deserializing certain values leads to infinite loop (#224)
We had an infinite loop where the `Value` tag invoke deserializer was
calling the `tl::expected<Value, JsonError` version, which called back
to itself.

This was manifesting as a runtime infinite loop. 

It's not caught by contract tests since I believe the trigger is having
a list of variations that are themselves arrays, like:
```
"variations": [["a","b","c",123,true,"false"], [ .. other stuff ..], ... ]
```
The fix is to break the cycle and remove the incorrect implementation.