fix: catch exception if en_US.utf8-locale missing when parsing datetime headers (#251)
The event delivery subsystem inspects the HTTP headers on the event
endpoints whenever it posts a batch of events.

It parses out the date header, storing it for later use. When emitting
debug events, it is used to ensure that events stop emitting even if the
host's time is off - as a form of time synchronization.

The parsing code is quite ugly, and also wrong - it assumed that
`en_US.utf-8` would always be available.

This commit hoists the locale loading routine out of the individual
flush-workers and into the worker pool. If it can't load the locale, it
emits a single `warn` log at startup explaining the effect.