chore: detect where OpenSSL is installed on Windows runner (#256)
We have intermittent CI failures because OpenSSL is installed to either
`Program Files\OpenSSL-Win64` or `Program Files\OpenSSL`. I'm not
entirely sure why, but it seems to be luck of the draw on the runner.

This should detect where it was installed and pass it to CMake. 

We can't force the install location because that requires a licensed
chocolatey. Chocolatey also writes out the install location - for this
particular package - but we'd have to process its output. Instead, this
just tests the two known installation directories.