chore: remove server-side version from release-please manifest (#272)
This is an attempt to fix the pending release. 

Ideally, release please would detect that no release is necessary or
that a 0.1.0 initial release of the server side only is necessary.

Right now, it thinks that the `sse-client` needs a release, and it's
therefore bumping that and the server/client SDKs.