chore: limit commit-search-depth to 10 for server-side package (#274)
It appears release-please is looking far back in history when
determining the release version of the server package. During this
search, it also appears to be running into a SSE commit that didn't use
conventional commit formatting.

Now, it's unclear if this is connected - but it might be making
release-please think that an SSE bump is required.

My theory is that by limiting to 10 commits (not sure how low this could
be - 0?) we can eliminate the issue. This should be ok since we know for
a fact no conventional commits have happened to SSE client that would
affect the server.