feat: refactor existing components into data system [2] (#291)
This PR moves the contents of `src/data_sources` (polling + streaming
implementations) into dedicated subdirectories of a new directory,
`data_components`.

The thousand-yard view is that we now have a larger abstraction (Data
System) of which these components are a part of.

The objects in `data_components` can be composed by SDK authors to
create the Background Sync and Lazy Load data systems. Background Sync
is the equivalent of streaming/polling, while Lazy Load operates against
a database.