refactor: rename IPush and IPull sources to better reflect their purposes (#295)
The original names were an attempting to retain some symmetry, while
also showing that both modes are "sources".

This doesn't need to be the case. These "sources" are conceptual, but
have totally different modes of interaction.

I've updated the names to `IDataSynchronizer` and `IDataReader`. These
are names that I used in past revisions of the DS spec, and I think they
are the better choice today.