refactor: improve IDataSynchronizer's interface to expose IDestination (#306)
This refactors `IDataSynchronizer` to take an `IDestination` and
bootstrap data in its `StartAsync` method, rather than having an
independent `Init` function.

The addition of `IDestination` makes the interface's purpose clearer.

Before the destination was hidden - you'd only see it when constructing
an implementation. Now it's clear that the `IDataSynchronizer` puts data
into the `IDestination`; no hidden dependency.

It also doesn't make sense to "Init" an `IDataSynchronizer` with an
SDKDataSet.

The purpose of that call is to allow the synchronizer to inspect any
existing data in possession by the SDK. The result of that inspection
might mean an optimized fetch of initial data is possible (e.g. if there
was an etag present in the initial data.)

---------

Co-authored-by: Matthew M. Keeler <keelerm84@gmail.com>