feat: add Shutdown method to IDataSystem (#307)
This adds a `Shutdown` method to `IDataSystem`, which allows the system
to stop any long-running asynchronous operations (such as
streaming/polling mode.)

Currently for the LazyLoad system, this is a no-op. In the future if we
have some sort of "active cache eviction" on an interval, this would be
used.