feat: omit empty items from data model JSON serialization (#309)
The server-side SDK is able to treat missing JSON fields as empty ones
in-memory.

For example, in a segment if `included` was missing then the SDK infers
it's an empty array.

This change makes the SDK write out data using that same logic. If an
in-memory field is conceptually empty, then omit it from the written
data.

This is an optimization, so the only places updated in this commit are top-level flag segment models which are frequently used in unit tests where we'd like to omit missing fields.