build: dynamically link boost when producing shared libraries (#284)
Boost 1.83 introduced a change where `boost.url/json` needs to be linked in rather than header-only.

We had a goal of hiding boost when building our shared libraries.
Previously, we did this by including boost.json and boost.url headers in
`boost.cpp`.

The need to link in these libraries poses a problem when we produce our shared library - the static ones we fetch for Linux weren't compiled as position-independent. 

This changes the default build scheme to link boost dynamically by default when compiling the SDK as a shared library. This means applications will need to ensure the shared boost libraries are present on the target system, or shipped with the application. 

A new `LD_DYNAMIC_LINK_BOOST` flag (on by default) is provided to revert this behavior and find static boost libs instead when compiling the SDK as a shared library.