fix: simulate mac + windows releases, fix openssl root path (#329)
Removes some obsolete openSSL CI config, as well as enables release
simulation for Mac + Windows (previously only on Ubunutu.)

This fixes Windows releases by setting `LD_DYNAMIC_LINK_BOOST=OFF` when
building shared libraries. This was already intended to be enabled, but
it was only being set on the debug releases (`build-windows.sh`). Now it
is present in both the debug and non-debug releases.