refactor: tombstone deserialization (#330)
This cleans up some nasty tombstone deserialization code in the initial
server-side release. It swaps from an ad-hoc deserializer to one
compatible with `boost::value_to`.

I've also removed the `StorageItem` concept that I had introduced to
make it clear that a deserialized item was a tombstone or not. That's
already taken care of by `ItemDescriptor` - I just wanted to see if a
more explicit data structure could work. I think it's not worth the
trouble at the moment.