fix: handle service endpoints with custom ports correctly (#389)
Users can configure the SDK with custom service endpoints. There was a
bug in the endpoint configuration process where using a custom port
would mean you couldn't also specify a HTTPS connection.

That is, if you specified something like `https://localhost:9090` for
testing, it would see `9090` and not setup the proper SSL context
needed. This had no affect on default usage of the SDK, because the
default LaunchDarkly endpoints don't specify a port.