feat: specify a custom CA file for TLS peer verification (#409)
This adds a new config builder option, `CustomCAFile` and associated C
binding to the server and client SDKs.

When specified, the SDK's streaming, polling, and event connections will
verify its TLS peer based on the CAs found in this file. The custom file
may be un-set by passing an empty string.