build: add LD_BUILD_EXPORT_ALL_SYMBOLS option to export all symbols (#420)
Address the issue https://github.com/launchdarkly/cpp-sdks/issues/411.

Adds a new CMake option, LD_BUILD_EXPORT_ALL_SYMBOLS, which exports all symbols,
including CPP and C APIs. It is set to OFF by default because the CPP ABI is not guaranteed to be stable.

This should only be used with caution if the requirements and consequences are fully understood. 

---------

Co-authored-by: Xiaofeng Feng <Xiaofeng.feng@autodesk.com>
Co-authored-by: Casey Waldren <cwaldren@launchdarkly.com>