feat: server-side payload filters (#435)
This PR introduces Payload Filters to the C++ Server-Side SDK. The
filter key can be configured individually on the Streaming or Polling
data sources.

If an invalid key is provided (empty or not matching regex), a runtime
error will be logged at the time of data source construction. The SDK
will then request an unfiltered environment as a failsafe.