feat: add client-side C binding for fetching data source state (#442)
We were missing a convenient way to convert a data source status's state
variable to a string. This is useful in logging and debugging
applications.

While adding this, I also applied the "ensure C enum is at least 32 bits
wide" treatment that has been done in the past on the LogLevel enum.