feat: preliminary find_package support (#452)
This PR adds preliminary support for:
```cmake

find_package(launchdarkly REQUIRED)

target_link_libraries(app 
        PRIVATE 
        launchdarkly::{launchdarkly-cpp-server,launchdarkly-cpp-client}
)
```
It treats the C++ monorepo as the installable unit. In the very near
future, this will likely be refactored so that we have two packages. If
you want both on the same host, you'd need to install both (and there'd
be some redundancy.)

Currently there is only one version, `0.1.0`. This isn't updated by
release-please.

For these two reasons, I'm indicating that this is preliminary and
subject to change form.