feat: expose prerequisite relations in AllFlags API (#463)
Updates the `AllFlags()` API to gather and expose prerequisite
information for flags.

The prereqs are available via `State::Prerequisites()`, and are more
importantly serialized in the JSON representation of the bootstrap
payload.

The implementation strategy was to create a custom `IEventProcessor`
specifically for `AllFlags` usage. Previously, `AllFlags` invoked the
evaluator with a no-op `EventScope`, which made event creation a no-op
within the eval algorithm.

This change now passes in a `PrereqEventRecorder` which tracks the
top-level prerequisites of each flag.