chore: Downgrade GHA runner to ubuntu-22.04 (#480)
gcc 13 (included in ubuntu-latest) no longer transitively includes
`<cstdint>`. This resulted in a break in redis-plus-plus, which is fixed
in v1.3.8. But we can't upgrade to that due to another issue with the
cmake stuff. So for now, we are going to lower the version of the GHA
runner we use.