fix: Ensure that serialization of a variation or rollout uses the correct overload. (#483)
The was boost::json handles tag_invokes seems to have changed slightly
such that serialization of a `VariationOrRollout` may be handled by the
built-in handling for a std::variant instead of our specialized
override.

This PR introduces the use of context to ensure our specific
serialization is always used in this case.

https://www.boost.org/doc/libs/1_87_0/libs/json/doc/html/json/conversion/contextual_conversions.html